% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c_RMSEP}
\alias{c_RMSEP}
\title{Calculate RMSEP Statistic from PLS results}
\usage{
c_RMSEP(Y_actual, Y_pred, ncomp = NULL)
}
\arguments{
\item{Y_actual}{Matrix of observed response values}

\item{Y_pred}{Matrix/array of predicted response values}

\item{ncomp}{Optional integer specifying number of components to evaluate}
}
\value{
Numeric vector of RMSEP values, either single value or one per component
}
\description{
An internal function that computes the Root Mean Square Error of
Prediction (RMSEP) for partial least squares models. Used within Bio.VIP for
assessing prediction accuracy.
}
\details{
Calculates RMSEP as square root of mean squared prediction error.
When ncomp is specified, computes RMSEP for each component up to ncomp.
RMSEP provides measure of prediction accuracy in original response variable units.
}
\references{
Mevik, B. H., & Wehrens, R. (2007). The pls package: principal component and partial
least squares regression in R. Journal of Statistical Software, 18(2), 1-24.
}
\seealso{
\code{\link{Bio.VIP}} \code{\link{AABA}}
}
\author{
Keeling et al., 2025
}
\keyword{internal}
