% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAV.R
\name{CAV}
\alias{CAV}
\title{CAV: Cluster Analysis Visualizations}
\usage{
CAV(
  Data,
  Group = NULL,
  ID = NULL,
  nboot = 1000,
  Dir_name = NULL,
  parallel = TRUE,
  method = "mahalanobis",
  ...
)
}
\arguments{
\item{Data}{a data table or data frame object.}

\item{Group}{a list object containing a vector of grouping variables.}

\item{ID}{an optional vector containing the observation identifications.}

\item{nboot}{the number of bootstraps to be run for the agglomerative cluster analysis.}

\item{Dir_name}{an optional character string indicating the name of a folder
either to be created or already exists in the working directory where generated
files can be stored.}

\item{parallel}{a boolean value indicating whether to use parallel processing.
Parallel processing is strongly recommended in cluster analysis.}

\item{method}{a character string indicating the distance matrix method that
should be applied on the data. Options include: "manhattan",
"euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower",
"altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao",
"mahalanobis", "chisq", "chord", "hellinger", "aitchison", or "robust.aitchison".
Please see the \emph{vegdist} "method" argument from the \strong{vegan} package for more
details.}

\item{...}{Optional arguments to be passed into the Predictor arguments for the
\code{DT} function. This is used to create distance matrices for the analysis.}
}
\value{
a list object of the following:
\itemize{
\item{k_medoid:   a pam object containing the results of the pam cluster analysis.}
\item{k_medoid_plot:    an integer indicating the number of k-medoid clusters selected.}
\item{dendro:   a phylo object containing the dendrogram from the agglomerative cluster analysis.}
\item{agg_plot:   a list object containing fan plots of the agglomerative clusters for each group}
}
}
\description{
This function conducts both an agglomerative hierarchical
cluster analysis and an k-medoids cluster analysis, generating
visually interactive figures to evaluate relationships among observations. Both
of these cluster analyses are unsupervised, that is, they aren't given group labels,
and instead use distance matrices to identify clusters of individual observations.
This is useful in contexts, such as geometric morphometrics, in identifying
taxonomic or even inter/intra-population differences based on their relationships
to one another.
}
\details{
This function accepts a data table and from it generates a distance matrix.
A variety of standardized methods for distance matrix calculation are avaialble
including mahalanobis distances. For a full list, please view the method argument
from the \emph{vegdist} function in the \strong{vegan} package. Mahalanobis distances
are calculated based on the \emph{mahalanobis} function from the \strong{stats} package.
Subsequently, this distance matrix is subject to a bootstrapped, agglomerative
hierarchical cluster analysis via the \emph{pvclust} function from the \strong{pvclust} package.
The agglomerative cluster analysis linkage method for bottom-up clustering
is automatically identified using the \emph{agnes} function from the \strong{cluster}
package by computing yhe agglomerative coefficient (AC), as a measure how well
the given data can be clustered for each possible method ("average", "single",
"complete", "ward", "weighted"). The method with the highest AC is used. Additionally,
a partitioning cluster analysis is conducted through k-medoids using the \emph{pam} function
from the \strong{cluster} package. Partitioning Around Medoids (PAM), as opposed to
k-means clustering, is more robust to data outliers and uses medoids in its clusterings.
First, the number of possible k clusters are iteratively performed using the \emph{pam}
function. Then the average silhouette width, an average measure of how well the data
fits the cluster, is calculated for each k and the k clusters with the highest
average silhouette width is selected as k. Then, the function is run again with
the selected parameters.

Furthermore, this function publishes each cluster plot into an automatically
generated folder whose directory name can be defined through Dir_name. Cluster
data from both the agglomerative and pam clustering is output from the function.
}
\examples{

\dontrun{
# Example grouping mandibular corpus landmarks by Collection (for agglomerative clustering)

# Import the sample 3D p x k x n array *Corpus_Land*
data(Corpus_Land)

# conduct a generalized procrustes analysis of the data

gpa_data = gpagen(Corpus_Land)

# Transform the 3D array as a data table using the geomorph package
table_corpus_land = data.frame(two.d.array(gpa_data))

# Import grouping variable Collection from the Biomechanics dataset
data(Biomechanics)
Group = Biomechanics[1:119,]$Collection
ID = unique(Biomechanics$ID)

# Run the cluster analysis vizualization tool

output = CAV(Data = table_corpus_land,
             Group = Group,
             ID = ID,
             nboot = 1000,
             Dir_name = "Preferred Directory Folder Name",
             parallel = TRUE,
             method = "mahalanobis"
             )

output
}

}
\seealso{
[\code{pam} \code{pvclust} \code{agnes}
}
\author{
Keeling et al., 2025
}
