% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_segments}
\alias{create_segments}
\title{Create a k-fold segmentation}
\usage{
create_segments(data, k = 5)
}
\arguments{
\item{data}{a data table or data frame for k-fold segmentation}

\item{k}{a numeric indice indicating how many folds should be performed.}
}
\description{
Creates a list of row indices for a dataset by which a k-fold
segmentation is performed.
}
\keyword{internal}
