% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{CSGM-summaries}
\alias{CSGM-summaries}
\title{Summarize Results for CSGM Package functions}
\arguments{
\item{object}{An object resulting from functions within the package.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
A collection of S3 methods used to extract and summarize the results from
regression based functions within the CSGM package.
}
\seealso{
\code{\link{AABA}} \code{\link{GMA}} \code{\link{reg.bio}}
}
\author{
Keeling et al., 2026
}
