% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT.R
\name{DT}
\alias{DT}
\title{Data Transformation (DT) - A function to transform data}
\usage{
DT(
  Response = NULL,
  Predictor = NULL,
  Res_transform = "NULL",
  Pred_transform = "NULL",
  Res_ncomp = NULL,
  Pred_ncomp = NULL,
  Res_point_set = NULL,
  Pred_point_set = NULL,
  dt_method = "procdist"
)
}
\arguments{
\item{Response}{an array, matrix, or data table for the response(dependent) variables.}

\item{Predictor}{an array, matrix, or data table for the predictor(independent) variables.

#' @param Res_transform Character string indicating which data transformation of the response
(i.e., landmark array) variable is desired. There are several acceptable options for data
transformation of the response variable:
\itemize{
\item procdist # Procrustes or other distances matrix
\itemize{
\item Calculates Procrustes or other distances as sqrt(sum(squared differences))
\item Accepts array, matrix, or data table
}
\item pca # Principal Component Analysis
\itemize{
\item Direct PCA on array, matrix, or data table
}
\item procdist+pca # Sequential Procrustes or other distances + PCA
\itemize{
\item Calculates Procrustes or other distances
\item PCA on resulting distance matrix
}
\item GM # Geometric Mean
\itemize{
\item Calculates geometric mean of all variables as exp(mean(log(x + 1e-8)))
\item For arrays: calculates landmark configuration centroid
}
\item procdist+GM # Sequential Procrustes or other distances + Geometric Mean
\itemize{
\item Calculates Procrustes or other distances
\item Geometric mean on resulting distances
}
\item 2D # Dimension reduction
\itemize{
\item Transforms 3D array to data frame
}
\item NULL # No transformation
\itemize{
\item Data remains unchanged
}
}}

\item{Pred_transform}{Character string indicating which data transformation of the predictor
(i.e., landmark array) variable is desired. There are several acceptable options for data
transformation of the response variable:
\itemize{
\item procdist # Procrustes or other distances matrix
\itemize{
\item Calculates Procrustes or other distances as sqrt(sum(squared differences))
\item Accepts array, matrix, or data table
}
\item pca # Principal Component Analysis
\itemize{
\item Direct PCA on array, matrix, or data table
}
\item procdist+pca # Sequential Procrustes or other distances + PCA
\itemize{
\item Calculates Procrustes or other distances
\item PCA on resulting distance matrix
}
\item GM # Geometric Mean
\itemize{
\item Calculates geometric mean of all variables as exp(mean(log(x + 1e-8)))
\item For arrays: calculates landmark configuration centroid
}
\item procdist+GM # Sequential Procrustes or other distances + Geometric Mean
\itemize{
\item Calculates Procrustes or others distances
\item Geometric mean on resulting distances
}
\item 2D # Dimension reduction
\itemize{
\item Transforms 3D array to data frame
}
\item NULL # No transformation
\itemize{
\item Data remains unchanged
}
}}

\item{Res_ncomp}{Optional integer. Number of components to retain for response variables.
Only applicable when Res_transform string includes PCA transformation. Default is NULL.}

\item{Pred_ncomp}{Optional integer. Number of components to retain for predictor variables.
Only applicable when Pred_transform string includes PCA transformation. Default is NULL.}

\item{Res_point_set}{an optional vector containing the numeric representation of variables
in the data table, matrix, or array to subset the response variables for this
data tranformation. If a subset is provided, the function will only perform
the data transformation on the selected variables.}

\item{Pred_point_set}{an optional vector containing the numeric representation of variables
in the data table, matrix, or array to subset the predictor variables for this
data tranformation. If a subset is provided, the function will only perform
the data transformation on the selected variables.}

\item{dt_method}{an optional character string used only when data transformation
for the response and/or predictor variable is set to an option with "procdist".
The default is "procdist". If default is changed, the function will overwrite the
procdist iteration and perform a distance matrix from the following methods:
"manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard",
"gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao",
"cao", "mahalanobis", "chisq", "chord", "hellinger", "aitchison", or
"robust.aitchison". Please see the \emph{vegdist} "method" argument from the
\strong{vegan} package for more details on distance matrix calculations.}

\item{subset}{an optional logical indicating whether only a subset of the variables
in either the response and/or predictor variables should be considered in the
data transformation.}
}
\value{
a nested list object with the following structure:

\itemize{
\item{Response:   Optional list object with transformed data (Response). If a variant of "pca" is selected, several objects will be included: Res_dim (original data dimensions), Res_p (number of original variables), Res_pca (principal component analysis object), Res_rot (the rotation matrix of the PCA), Res_pca_scores (the principal component scores). Used in other CSGM functions}

\item{Predictor:  Optional list object with transformed data (Predictor).}
}
}
\description{
This function takes an array, matrix, or data table object and
transforms the data into either a distance matrix, principal components,
geometric mean, or 2D data table. This function can also normalize the data
and allows for subsetting of data.
}
\details{
This is an internal function was designed to transform variables prior to performing
a regression. It accepts either a p x k x n (variables, dimensions, observations)
array, a matrix, or data table object as the Response and/or Predictor variables.
This function allows for data transformation of subsetted data by switching
the argument "subset" to TRUE and providing a numeric vector of the numeric
representations (i.e., column number) of the variables of interest in the data
object. This function provides a number of data transformation types such as:
"procdist" (procrustes or another distance matrix), "procdist+pca" (Procrustes
or another distance matrix then subjected to principal component analysis),
"procdist+GM" (Procrustes or another distance matrix then the Geometric Mean
is calculated), "pca" (Principal Component Analysis), "GM" (Geometric Mean),
2D (Change a 3D array to a 2D data table), NULL (do not transform data). Each
of these options are common methods to transform geometric morphometric or
biological variables for downstream analysis. If 'procdist', or any option with
'procdist', is selected the method argument will be used to select which
distance matrix calculation will be used to trasnform the data. The option
'procdist' or Procrustes distances will be selected by default unless changed.
Procrustes distances are calculated as the square root sum of all squared distances.
All other distance matrix calculations use the \emph{vegdist} function from the
\strong{vegan} package. Principal Component Analysis uses the \emph{gm.prcomp} function
from the \strong{geomorph} package for 3D arrays. For 2D data tables, the \emph{PCA}
function from \strong{factoextra} is applied.
}
\examples{
\dontrun{
# Scenario: Create a Procrustes distance matrix of a landmark configuration
# and find the geometric mean for all corresponding biomechanical variables.

# Import data sample of mandibular corpus: Corpus_Land
# Import data sample of mandibular corpus data table: Biomechanics

data(Corpus_Land)
data(Biomechanics)

# Conduct the data transformation

# NOTE: only numeric values can be taken for geometric means for the biomechanics dataset
# NOTE: Use subset for Biomechanic data table to select only numeric values

output <- DT(Response = Corpus_Land,
             Predictor = Biomechanics,
             Res_transformation = "procdist",
             Pred_transformation = "GM",
             subset = TRUE,
             point_set_res = NULL, # The response variable is not subset.
             point_set_pred = c(5:93) # only numeric variables
             ) # all other arguments leave as default

Land_matrix = output$Response # our Procrustes distances

GM_Bio = output$Predictor # our Geometric Means of all biomechanical variables

# Add identifying information back into GM_Bio

GM_Bio = data.frame(Biomechanics[,1:4], GM_Bio)

# Since the Biomechanic data has 5 cross-sections and duplicated IDs, we need to
# restructure the data so each cross-section is a variable of a geometric mean
# for each individual.

GM_Bio = GM_Bio \%>\% summarise('ID' = unique(GM_Bio[,1]),
                              'LM1M2' = GM_Bio[,5] \%>\% filter(Region == "LM1M2"),
                              'LP3P4' = GM_Bio[,5] \%>\% filter(Region == "LP3P4"),
                              'Symphysis' = GM_Bio[,5] \%>\% filter(Region == "Symphysis"),
                              'RP3P4' = GM_Bio[,5] \%>\% filter(Region == "RP3P4"),
                              'RM1M2' = GM_Bio[,5] \%>\% filter(Region == "RM1M2")
                             )
# view the data

view(Land_matrix)

view(GM_Bio)
}
}
\seealso{
\code{gm.prcomp} \code{PCA}
}
\author{
Keeling et al., 2025
}
