% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex_data.R
\name{Ex_data}
\alias{Ex_data}
\title{Ex_data}
\usage{
Ex_data(Directory, Land)
}
\arguments{
\item{Directory}{a file list indicating the directory that the neseted folder should be created}

\item{Land}{a p x k x n (landmarks, dimensions, observations) array of a landmark configuration}
}
\value{
a nested list object in a user-specified directory called "Example"
containing folders with their corresponding landmark formats.
}
\description{
This script takes a 3D landmark array and generates raw landmark files in various formats
commonly used in geometric morphometrics. The data structure is preserved across
all formats.
}
\examples{
\dontrun{
data(Corpus_Land)

Directory <- "E:/Paper 2 - The CSGM Package/CSGM/data/Example"

Ex_data(Directory, Corpus_Land)

}
}
\author{
Keeling et al., 2025
}
