% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_land.R
\name{extract_id}
\alias{extract_id}
\title{Extract IDs from filenames (handles both numeric and alphanumeric IDs)}
\usage{
extract_id(filepath)
}
\description{
Extract IDs from filenames (handles both numeric and alphanumeric IDs)
}
\keyword{internal}
