% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_curves.R
\name{gen_curves}
\alias{gen_curves}
\title{Generate a curve matrix for landmark data}
\usage{
gen_curves(ncurves, nsemiland, def_semi, def_points, closed = FALSE)
}
\arguments{
\item{ncurves}{Numeric. Total number of distinct curves (e.g., 10).}

\item{nsemiland}{Numeric. Total number of landmarks in the dataset (e.g., 590).}

\item{def_semi}{Numeric vector. Indices of the semi-landmarks.}

\item{def_points}{Numeric vector. Indices of the fixed landmarks (anchors).}

\item{closed}{Logical. Whether INDIVIDUAL curves are closed loops.}
}
\value{
This function returns a n x 3 landmark curve matrix where n is
dependent on the number of landmarks and semilandmarks in the
landmark configuration.
}
\description{
This function generates a 'n x 3' curve matrix containing the positions of
semilandmarks relative to fixed landmarks. This curve matrix can be
subsequently used to calculate equidistant semilandmark or landmark points.
This function also satisfies the "gpagen" function's argument 'curves' from
the \emph{geomorph} package.
}
\details{
This function is primarily for landmark and semilandmark
configurations. A n x 3 landmark curve matrix is generated as a result which
can be used as an argument for functions to manage landmark spacing. This
function is only recommended for situations when the landmark configuration
contains (1) only semi-landmarks whose spacing needs further management,
(2) only landmarks that need to be equidistantly spaced, or (3) the landmark
configuration contains both semilandmarks and landmarks.
}
\examples{
\dontrun{
# You should know the exact positions of your semilandmarks in your landmark configuration before using this function
data(semiland_indices)
data(anchor_indices)
gen_curves(ncurves = 10,
           nsemiland = 590,
           def_semi = semiland_indices,
           def_points = anchor_indices,
           closed = FALSE)
}
}
\seealso{
\pkg{\link{geomorph}}
}
\author{
Keeling et al., 2026
}
