% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extreme_shapes.R
\name{get_extreme_shapes}
\alias{get_extreme_shapes}
\title{Generate Extreme Principal Component Shapes}
\usage{
get_extreme_shapes(Data_PCA, k, PCs)
}
\arguments{
\item{Data_PCA}{A PCA object which contains the following elements:
x (principal component scores), sdev (eigenvalues), rotation (eigenvectors),
center (2D matrix of mean shape of individuals within the PCA).}

\item{k}{An integer value indicating the number of dimensions in the
landmark configuration. Inputs of 2 or 3 are accepted.}

\item{PCs}{An integer value indcating the number of meaningful
Principal Components within the PCA by which the function should
calculate mean shapes for.}
}
\value{
a list object containing the following:

\itemize{
\item {negative PC shape for PC 1}
\item {positive PC shape for PC 1}
\item {negative PC shape for PC i}
\item {positive PC shape for PC i}
\item {... PCs alternate continues until all interested PCs are estimated}
}
}
\description{
This function generates a list containing the mean, positive, and negative
shapes for each Principal Component of a Principal Component Analysis.
}
\details{
This function takes a list object containing several important variables
commonly calculated during principal component analysis
(i.e., principal component scores, eigenvalues and eigenvectors,
number of interested principal components, number of dimensions)
and estimates the landmark configurations for the minimum and maximum observations.
Data_PCA requires four components supplied by the \emph{gm.prcomp} function from the
\strong{Geomorph} package:
\itemize{
\item{x           principal component scores}
\item{sdev        eigenvalues}
\item{rotation    eigenvectors}
\item{center      consensus (center of PCA) shape as a 2D matrix}
}
}
\examples{

\dontrun{

library(geomorph)
# Import sample dataset human_corpus

data(human_corpus)

Data_PCA = gm.prcomp(human_corpus)

output = get_extreme_shapes(Data_PCA,
                            k = 3,
                            PCs = 2)

# view the shape extremes

lolshape(output[[[1]]])

}
}
\seealso{
\code{\link{gm.prcomp}}
}
\author{
Keeling et al., 2025
}
