% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vignette.R
\name{get_vignette}
\alias{get_vignette}
\title{Extract Package Vignette to conduct sample analysis}
\usage{
get_vignette(vignette_name = "CSGM-Workflow", dest_dir = getwd())
}
\arguments{
\item{vignette_name}{the name in quotations of the vignette you want to extract}

\item{dest_dir}{the directory destination you want the file to be stored}
}
\value{
a .Rmd RMarkdown file of the package vignette in your project directory
}
\description{
This function finds the package's vignette and extracts it as
an .Rmd RMarkdown file to be used in the script panel. The file will appear in
the project directory.
}
\author{
Keeling et al., 2025
}
