% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hullgen.R
\name{hullgen}
\alias{hullgen}
\title{hullgen: an interactive group convex hull generator for Principal Component Analysis}
\usage{
hullgen(
  Data_PCs,
  select_PC,
  Group,
  color = NULL,
  ID = NULL,
  PCA_var = NULL,
  sym = FALSE
)
}
\arguments{
\item{Data_PCs}{a data table containing principal component scores for each observation.}

\item{select_PC}{a vector indicating which principal components are to be compared.}

\item{Group}{a vector containing the grouping values for each individual.}

\item{color}{a vector containing the color values for each group.}

\item{ID}{an optional vector indicating the individual names for each specimen in the sample}

\item{PCA_var}{an optional data table containing three columns: the principal
component number, the explained percentage of variance for each principal component,
and the cumulative explained percentage of variance. This is automatically
produced by the \emph{pca_variances} function.}

\item{sym}{a boolean indicating whether groups should take on different symbols
in the interactive plot.}
}
\value{
a plotly object containing a plot with group convex hulls.
}
\description{
This function generates a convex hull to visualize group dynamics in a
principal component, or similar, space. hullgen utilizes the workhorse
package "plotly" to generate interactive plots with the addition of 3D and 2D convex hulls.
}
\details{
This function generates an interactive convex hull plot to highlight group
differences between variables or principal components. This plot uses the
\emph{plotly} function from the \strong{plotly} package to generate the visualization.
To generate the indices used to construct the convex hull, this function uses
the \emph{chull} function from the \strong{grDevices} package for 2D convex hulls and the
\emph{convhulln} function from the \strong{geometry} package for 3D convex hulls.
}
\examples{

\dontrun{

# Example 1: Convex hull generation of principal component data from mandibular corpus shape

# Import the sample 3D p x k x n array *Corpus_Land*
data(Corpus_Land)

# Import grouping variable Collection from the Biomechanics dataset
data(Biomechanics)
Group = Biomechanics[1:119,]$Collection
ID = unique(Biomechanics$ID)

# Conduct a Principal Component Analysis on the 3D array using *gm.prcomp* from **geomorph**
Data_PCA = gm.prcomp(Corpus_Land)
Data_PCs = Data_PCA$x # extract only principal component scores

# Summarize the results with the pca_variances function
PCA_var = pca_variances(Data_PCA) # input object from the PCA

# Generate a PCA plot of PCs 1-2 with a convex hull based on Collection
output = hullgen(Data_PCs,
                 Group,
                 color = NULL, # allow color autofill for each group
                 ID,
                 PCA_var,
                 sym = FALSE # make all groups share the same symbol (circle)
                 )

# visualize the plot
output

# Save the plot
saveWidget(output, file = "output.html")
}
}
\seealso{
\code{plotly} \code{chull} \code{convhulln}
}
\author{
Keeling et al., 2025
}
