% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolligen.R
\name{lolligen}
\alias{lolligen}
\title{lolligen: An Interactive Principal Component Analysis Visualization Tool}
\usage{
lolligen(
  Data_PCA,
  k = 3,
  PCA_var = NULL,
  ID = NULL,
  coord_name = NULL,
  Group = NULL,
  no.shape = FALSE,
  subset = NULL
)
}
\arguments{
\item{Data_PCA}{a list object resembling or formed by the \emph{pca_variances} function output.}

\item{k}{an integer indicating the number of dimensions in the landmark configuration
prior to data transformation via PCA. k = 1 is for non-landmark data. Default value is 3.}

\item{PCA_var}{an object from the \emph{pca_variances} function output.}

\item{ID}{a vector containing the identification names and/or numbers for each
observation from the principal component analysis.}

\item{coord_name}{a vector listing the names of the landmark coordinates.}

\item{Group}{a list object containing one or more grouping variable vectors.}

\item{no.shape}{a boolean value indicating whether the function should
estimate and save the shape extremes of each important principal component.
Default value is set to FALSE. If set to TRUE, the function will not generate
interactive plots to visualize landmark variance based on each principal component.}

\item{subset}{an optional vector listing a grouping variable for each landmark.
If subset is added and no.shape is set to FALSE, the landmark lollipop graph
will segment each of the landmarks by the subset grouping, allowing the user
to show and hide groups of landmarks to visualize specific shape variances.}
}
\value{
a list object containing the following:

\itemize{
\item \emph{PCA_list}:  a list object containing all possible principal component plot pairs.
\item \emph{PC_mean_max_min}:   a list object containing the mean, maximum extreme,
and minimum extreme shapes or combination of variables respectively for each
meaningful principal component.
\item \emph{lollipop_PCs}:  a list containing a series of lollipop graph plots
for each PC. Each plot already contains the mean, maximum, and minimum shape
extremes to visualize the shape variations across each principal component.
Keep in mind that double brackets are required to call each PC (see examples).
\item \emph{PCA_Group}:   an optional list object containing the names of each grouping
variable supplied through the Group argument. If NULL, this object will
not be generated. If argument Group is not NULL, a convex hull plot will be
generated for each principal component pairing for each group vector in the list.
}
}
\description{
This function takes a principal component analysis (PCA) object via the
internal function pca_variances and generates several interactive visual aids
commonly required to analyze interindividual and multivariate variance in PCA.
This function primarily uses the \strong{plotly} package to generate interactive
PCA plots between principal components as well as interactive lollipop graphs
showing the extreme shape variances for each meaningful principal component.
}
\details{
This function requires a list object generated by the \emph{gm.prcomp} function in
the \emph{geomorph} package, or a list containing the same variables (i.e., principal
component scores, eigenvalues and eigenvectors, number of interested principal components,
number of dimensions). The function then uses the argument \emph{PCA_var}, an output
from the \emph{pca_variances} function, and generates automatically labeled,
interactive principal component plots that can be used to analyze
interindividual variation and also evaluate variable relationships through lollipop
graphs of the variables used to conduct the PCA. This function will accept
any list object for Data_PCA as long as it contains the following objects:
\itemize{
\item \emph{x}:   a data table of the principal component scores (columns) and observations (rows)
\item \emph{sdev}:  a data table containing the eigenvalues from the PCA.
\item \emph{rotation}:  a data table containing the rotation matrix or eigenvectors
calculated during the PCA.
\item \emph{center}:  a 2D or 3D p x k x n array containing the mean variable configuration
}
Additionally, lolligen allows a list object containing vectors of group labels
for each observation in the landmark configuration. This function does not
conduct a between group PCA. If this is desired, Data_PCA should contain the values
output by a between groups PCA such as the \emph{bgPCA} function from the \strong{ade4} package.
From the data presented in Data_PCA, the function will generate convex hulls and color
the individuals by each grouping variable in the list and store each plot separately.
}
\examples{

\dontrun{

# Example 1: Without groupings or subsets

# Import the sample 3D p x k x n array *Corpus_Land*
data(Corpus_Land)

# Conduct a Principal Component Analysis on the 3D array using *gm.prcomp* from **geomorph**
Data_PCA = gm.prcomp(Corpus_Land)

# Summarize the results with the pca_variances function
PCA_var = pca_variances(Data_PCA) # input object from the PCA

# View the scree plot
PCA_var$Scree

# View the number of meaningful PCs
PCA_var$PCs

# Conduct the lolligen function and input the number of desired principal components.

output <- lolligen(Data_PCA, # PCA object
                   k = 3, # three dimensional landmark data
                   PCA_var # output from the internal pca_variances function
                   )

# Visualize individual variation between principal components 1 and 2
output$PCA_list$PC_1_2

# Visualize

# Visualize shape changes along PC 1

output$lollipop_PCs[[1]]

# Visualize shape changes along PC 2

output$lollipop_PCs[[2]]

# Save the plot

saveWidget(output$lollipop_PCs[[1]], "PC_1_Shape_Changes.html")

# Example 2: lolligen with groupings and subsetted landmarks

# Import the sample dataset Corpus_Land
data(Corpus_Land)

# Import the associated sample Biomechanical data for IDs and grouping variables.
data(Biomechanics)[1:119,1:3] # we are only interested in the groupings.

# Conduct a Principal Component Analysis on the 3D array.
Data_PCA = gm.prcomp(Corpus_Land)

# Summarize the results with the pca_variances function
PCA_var = pca_variances(Data_PCA) # input object from the PCA

# Create a list containing randomized Grouping variables

Group = list(Collection = Biomechanics$Collection)
ID = unique(Biomechanics$ID)

# Conduct the lolligen function and input the number of desired principal components.

output <- lolligen(Data_PCA, # PCA object
                   k = 3, # three dimensional landmark data
                   PCA_var, # output from the internal pca_variances function
                   ID = ID, # specimen IDs
                   coord_name = dimnames(Corpus_Land)[[[1]]], # landmark names
                   Group = Group, # list of grouping variables
                   no.shape = FALSE, # no mesh generation is required
                   subset = c(rep("LM1M2", "LP3P4", "Symphysis", "RP3P4", "RM1M2"), each = 120)
                   )

# Visualize individual variation between principal components 1 and 2
output$PCA_list$PC_1_2

# Visualize grouped PCA plot between principal components 1 and 2

output$PCA_Group$Collection$PC_1_2

# Visualize shape changes along PC 1

output$lollipop_PCs[[1]]

# Visualize shape changes along PC 2

output$lollipop_PCs[[2]]

}
}
\seealso{
\code{plotly} \code{hullgen}
}
\author{
Brian Anthony Keeling
}
