% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolshape.R
\name{lolshape}
\alias{lolshape}
\title{lolshape - Am interactive tool to Visualize shape data}
\usage{
lolshape(
  shape.list,
  ID = NULL,
  title = NULL,
  coord_name = NULL,
  color = NULL,
  subset = NULL,
  mesh = FALSE
)
}
\arguments{
\item{shape.list}{a list object containing p x k x n arrays and/or 3D meshes to be visualized.}

\item{ID}{an optional vector containing the identification for each observation in the landmarked sample.}

\item{title}{an optional vector to create a custom title for the interactive graphic.}

\item{coord_name}{a vector listing the names of the landmark coordinates.}

\item{color}{an optional vector containing the color of each observation in shape.list.}

\item{subset}{An optional vector specifying the subset or group names associated
with each landmark in the landmark configuration. If provided, the resulting
interactive figure will allow toggling between these landmark subsets.}

\item{mesh}{a logical value indicating whether shape.list contains at least
one mesh. If mesh is set to TRUE, the function will generate interactive visuals
for the meshes and landmark configurations separately and also as a single combined
interactive visualization.}
}
\value{
a list object containing the following:

\itemize{
\item \emph{Graph}:   a plotly object containing only the landmarks listed in shape.list.
\item \emph{Meshes}:  a plotly object containing an interactive visual for all of
the meshes listed in shape.list. Only applicable if mesh = TRUE.
\item \emph{Combined}:  a plotly object containing an interactive visual that
shows both 3D meshes and landmarks in the same shape space. Only applicable
if mesh = TRUE.
}
}
\description{
This function allows the interactive comparison of shape based variation between
two or more landmark configurations and/or meshes. Interactive graphics are made
using functions within the \pkg{plotly} package. This function is best for
comparing inter/intra-individual shape comparisons, mean group shape comparisons,
and compare shape predictions with the original. This function provides a versatile
graphic that permits the visibility of shape configurations and landmark subsetting.
}
\details{
This function requires a list of p x k x n arrays and/or 3D mesh objects. Both
arrays and meshes can be provided in this list but the arrays must be only either
2D or 3D. Function takes these shape arrays and/or meshes and transforms them
into a single, interactive \pkg{plotly} graphic to make interobservation comparisons.
If meshes are added, argument mesh must be set to TRUE. If mesh = TRUE, the function
will create three outputs: two to separate meshes from landmark arrays and another
which combines the two together. Optional arguments are added to allow for landmark
subsetting, object colors, plot title, and adding landmark names.
}
\examples{
\dontrun{

# Import the sample dataset Corpus_Land
data(Corpus_Land)

# Conduct a Generalized Procrustes Analysis
gpa_land = gpagen(Corpus_Land)

# Create visualization for the shape comparison between individuals 1212 and 6220
output = lolshape(shape.list = list('10' = landmarks[,,"10"],
                                    '20' = landmarks[,,"20"]),
                  ID = c('10', '20'),
                  title = "Shape Comparison Between individuals 10 and 20",
                  coord_name = dimnames(Corpus_Land)[[1]],
                  color = c("orange", "blue"),
                  subset = rep(c("LM1M2", "LP3P4", "Symphysis", "RP3P4", "RM1M2"), each = 120),
                  mesh = FALSE)
 # Visualize output
 output

 # Save interactive graphic to computer as html
 saveWidget(output, "output.htnl")

}
}
\seealso{
\code{\link[plotly:plotly]{plotly::plotly()}}
}
\author{
Keeling et al., 2025
}
