% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{m_R2}
\alias{m_R2}
\title{Multivariate Coefficient of Determination (R^2) Calculation}
\usage{
m_R2(observed, predicted)
}
\arguments{
\item{observed}{Matrix of observed values}

\item{predicted}{Matrix of predicted values}
}
\value{
a numeric value containing the coefficient of determination
}
\description{
Calculates appropriate R2 based on data characteristics:
\enumerate{
\item Correlation-based R2 for PCA/ranked data
\item Regularized trace-based R2 with Ledoit-Wolf shrinkage for high-dimensional data (p > n)
\item Trace-based R2 for standard multivariate data
\item Determinant based approach
}
}
\details{
A function to conduct
}
\keyword{internal}
