% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_variances.R
\name{PCA_variances}
\alias{PCA_variances}
\title{A Principal Component Analysis Statistics Compiler}
\usage{
PCA_variances(Data_PCA, PCs = TRUE)
}
\arguments{
\item{Data_PCA}{A PCA object which contains the following elements:
x (principal component scores), sdev (eigenvalues), rotation (eigenvectors),
center (mean shape of individuals within the PCA).}

\item{PCs}{An optional logical value indicating whether to calculate the number
of meaningful PCs. As this involves factor analysis, this may take some time to
calculate. The default is to perform Principal Component factor analysis.}
}
\value{
a list object containing the following:

\itemize{
\item \emph{pca_var}:   data table of dimensions pc x 3 where pc is the number of principal
components obtained from the prinicipal component analysis.
\item \emph{scree}:   principal component scree plot generated from the data table.
\item \emph{PCs}:   number of meaningful principal components.
}
}
\description{
This function takes the results of a principal component analysis object
and outputs standard statistics of the principal component analysis in a
table format. This function generates a table "PCA_var" which can be used
with the function 'Lolligen' which provides information on the variance
explained by each principal component as well as an analysis of the
meaningful number of Principal Components based on the data.
}
\details{
This function takes a list object containing several important variables
commonly calculated during principal component analysis
(i.e., principal component scores, eigenvalues and eigenvectors,
number of interested principal components, number of dimensions)
and generates several objects: a table containing the percentages of
explained variance for each principal components, a scree plot of the
explained variance by each principal component, and the number of meaningful
principal components. The meaningful number of principal components uses
factor analysis via the function \strong{fa.parallel} from the \emph{psych} package.
This function examines the percentages of variance across each principal
component in comparison to a random data matrix of the same size to assess
the most important number of principal components that can explain the
most significant percentages of variance. If the factor analysis is unable
to detect the most important number of principal components, by default the
number of principal components that can explain at least 80\% of the total
cumulative variance will be selected.
}
\examples{

\dontrun{

# Import sample dataset human_corpus
data(Corpus_Land)

# Conduct a Principal Component Analysis on the 3D array.
Data_PCA = gm.prcomp(Corpus_Land)

# execute the function
output = pca_variances(Data_PCA) # object from the PCA

# View the scree plot
output$Scree

# View the data table
output$PCA_var

# View the number of meaningful PCs
output$PCs

}
}
\author{
Keeling et al., 2025
}
