% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pcland}
\alias{pcland}
\title{Convert Principal Component Analysis Data to Original Data}
\usage{
pcland(
  Res_pca_scores = NULL,
  Res_rot = NULL,
  Res_pca_center = NULL,
  Res_ncomp = NULL,
  Res_p = NULL,
  Res_dim = NULL
)
}
\arguments{
\item{Res_pca_scores}{a matrix or data table containing principal component scores (eigenvalues).}

\item{Res_rot}{a matrix or data table containing eigenvectors (rotation matrix) from a principal component analysis.}

\item{Res_pca_center}{a 2D matrix of a landmark configuration representing the center shape across all principal components.}

\item{Res_ncomp}{an optional numeric vector containing the number of principal components used in a regression.}

\item{Res_p}{The number of landmarks of the original landmark configuration.}

\item{Res_dim}{The number of dimensions of the original landmark configuration.}
}
\value{
either 2D or 3D array of the predicted landmark configuration.
}
\description{
This function takes the eigenvalues, eigenvectors, and centered
principal component data to conduct a backwards reconstruction of the original
landmark configuration. This function is designed for reconstructing predicted
landmark configurations based on principal component data.
}
\details{
This function estimates a landmark configuration based on the
eigenvalues, eigenvectors, and center landmark configuration from a Principal
Component Analysis.
}
\examples{
\dontrun{

 # Load in Corpus_Land test data

 data(Corpus_Land)

 # Record the number of landmarks and dimensions

 Res_p = 600
 Res_dim = 3

 # Conduct a Principal Component Analysis using the DT function

 Response = DT(Response = Corpus_Land, Res_transform = "pca")

 # Extract the Eigenvalues, Eigenvectors, and Center Shape

 Res_pca_scores = Response$Res_pca_scores
 Res_rot = Response$Res_rot
 Res_pca_center = Response$Res_pca_center

 # Conduct the reconstruction

 output = pcland(Res_pca_scores,
                 Res_rot,
                 Res_pca_center,
                 Res_ncomp,
                 Res_p,
                 Res_dim)

}

}
\author{
Keeling et al., 2025
}
\keyword{internal}
