% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict.mixo_spls}
\alias{predict.mixo_spls}
\title{Predict Method for sPLS through MixOmics Package}
\usage{
\method{predict}{mixo_spls}(object, newdata, ...)
}
\arguments{
\item{object}{A mixo_spls object.}

\item{newdata}{Matrix or data frame of new predictor data.}

\item{...}{Additional arguments.}
}
\description{
S3 method to enable prediction for mixo_spls objects created by the standalone spls function.
This is required for plsr_stats to function correctly.
}
