% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pslr.bio.R
\name{pslr.bio}
\alias{pslr.bio}
\title{Procrustes Shape Linear Regression (pslr.bio)}
\usage{
pslr.bio(
  Models,
  paired_subsets = FALSE,
  point_set = NULL,
  all_vips = NULL,
  subset_vips = NULL,
  key = NULL,
  ...
)
}
\arguments{
\item{Models}{Nested list object containing regression hypotheses to test. Each list element should
contain Response and Predictor data, structured as:

\itemize{
\item 'Hypothesis Test Models' # Argument Models object name
\itemize{
\item 'Symphyseal shape ~ Posterior Corpus Properties' --- 1st hypothesis model
\itemize{
\item 'Symphyseal shape ~ LM1-M2 Corpus Properties ' --- 1st hypothesis test
\itemize{
\item Symphyseal Landmarks --- Response Data within hypothesis test
\item LM1-M2 Corpus Properties --- Predictor data within hypothesis test
\item dt_parameters --- Optional data transformation parameters
}
\item 'Symphyseal shape ~ LP3-P4 Corpus Properties ' --- 2nd hypothesis test
\itemize{
\item Symphyseal Landmarks --- Response Data within hypothesis test
\item LP3-P4 Corpus Properties --- Predictor data within hypothesis test
\item dt_parameters --- Optional data transformation parameters
}
\item 'Symphyseal shape ~ RP3-P4 Corpus Properties ' --- 3rd hypothesis test
\itemize{
\item Symphyseal Landmarks --- Response Data within hypothesis test
\item RP3-P4 Corpus Properties --- Predictor data within hypothesis test
\item dt_parameters --- Optional data transformation parameters
}
\item 'Symphyseal shape ~ RM1-M2 Corpus Properties ' --- 4th hypothesis test
\itemize{
\item Symphyseal Landmarks --- Response Data within hypothesis test
\item RM1-M2 Corpus Properties --- Predictor data within hypothesis test
\item dt_parameters --- Optional data transformation parameters
}
}
\item 'Symphyseal shape ~ Posterior Corpus Properties' --- 2nd hypothesis model
\itemize{
\item 'Symphyseal shape ~ All Corpus Properties ' --- 1st hypothesis test
\itemize{
\item Symphyseal Landmarks --- Response Data within hypothesis test
\item Posterior Corpus Properties --- Predictor data within hypothesis test
\item dt_parameters --- Optional data transformation parameters
}
}
\item ..... --- 3rd hypothesis model
}
}}

\item{paired_subsets}{a logical value indicating whether or not the function
should iterate through regressions for every possible combination of subsets.
If argument paired_subsets = TRUE and argument point_set is not NULL, then
the function will calculate all of the possible subset combinations between
the response and predictor variables and conduct separate regressions for
each subsetted combination.}

\item{point_set}{Optional nested list specifying variable subsets to analyze. Structure must
match Models argument with added subset definitions:
\itemize{
\item point_set
\itemize{
\item Model_1 # Argument object name
\itemize{
\item Response               # Response list subsets (e.g., landmarks)
\itemize{
\item Subset_1 = c(1:20)    # Landmarks 1-20
\item Subset_2 = c(21:40)   # Landmarks 21-40
\item Subset_3 = c(41:60)   # Landmarks 41-60
}
\item Predictor              # Predictor subsets (e.g., data table)
\itemize{
\item Subset_1 = c(1:5)     # Variables 1-5
\item Subset_2 = c(6:10)    # Variables 6-10
\item Subset_3 = c(11:15)   # Variables 11-15
}
}
}
}

Each subset should be a numeric vector specifying which landmarks/variables to include.
Subsets can be of different sizes and overlap. When paired_subsets=TRUE, all
possible combinations of response and predictor subsets will be analyzed.}

\item{all_vips}{Optional nested list object containing variables of interest to trim the predictor dataset.
The structure should follow a similar structure to the  point_set argument where
for each m hypothesis model, there are i hypothesis tests. Within each hypothesis
test should be a vector called "vip" of variables to include in the analysis.
\itemize{
\item all_vips
\itemize{
\item Model_1 (1st hypothesis model)
\itemize{
\item Test_1 = c("var1", "var2", "var3", ...) # Variables for 1st test
\item Test_2 = c("var1", "var4", "var8", ...) # Variables for 2nd test
\item ... # More tests
}
\item Model_2 (2nd hypothesis model)
\itemize{
\item Test_1 = c("var2", "var5", "var9", ...) # Variables for 1st test
\item ... # More tests
}
\item ... # More models
}
}}

\item{subset_vips}{Optional nested list object containing variables for each subset within hypothesis tests.
The structure should follow a similar structure to the  point_set argument where
for each m hypothesis model, there are i hypothesis tests. Within each hypothesis
test should be j list objects of character or numeric vectors called "vip" to
include in the analysis for each subset.
\itemize{
\item subset_vips
\itemize{
\item Model_1 (1st hypothesis model)
\itemize{
\item Test_1 (1st hypothesis test)
\itemize{
\item Subset_1 = c("var1", "var2", ...) # Variables for subset 1
\item Subset_2 = c("var3", "var5", ...) # Variables for subset 2
\item ... # More subsets
}
\item Test_2 (2nd hypothesis test)
\itemize{
\item Subset_1 = c("var2", "var4", ...) # Variables for subset 1
\item ... # More subsets
}
\item ... # More tests
}
\item Model_2 (2nd hypothesis model)
\itemize{
\item ... # Similar structure
}
\item ... # More models
}
}}

\item{key}{Character string. Identifier for output files and directories. Default is NULL.}
}
\value{
Returns a nested list containing:
\itemize{
\item{Model_1:   First Hypothesis Model
\itemize{ First Test within first Hypothesis Model
\item{ANOVA tables         Contains SS, MS, Rsq, F, Z, and P-values}
\item{Effect size          R-squared values for each model}
\item{Model coefficients   When applicable}
}
}
}

If write.results=TRUE, also generates Excel files containing:
\itemize{
\item Separate sheets for each regression analysis
\item Summary statistics and test results
\item Organized by model structure
}

Directory structure created:
\itemize{
\item pslr.bio_Results_(key)/Model_(model_name).xlsx
}
}
\description{
This function conducts, in parallel, a Procrustes linear regression
on multiple models between landmark arrays (Response variable) and predictor
variables (univariate or multivariate). The function was designed to aid in
studies where multiple regression hypotheses need to be tested on different
and/or the same predictor variables. This function permits conducting regressions
on subsets or combinations of subsetted landmarks and/or variables.
}
\details{
The pslr.bio function performs Procrustes shape linear regression analysis by:
\enumerate{
\item Creating working directories for results storage
\item Processing multiple regression models in parallel
\item Handling data transformations (PCA, Procrustes distances, etc.)
\item Supporting subset analysis of landmarks and variables
\item Implementing paired comparisons when requested
\item Managing error handling and alternative approaches
\item Writing results to Excel in summarized format
}

The function uses geomorph::procD.lm() for the regression analysis with:
\itemize{
\item Parallel processing enabled
\item 1000 iterations
\item Randomized residual permutation procedure (RRPP)
}
}
\examples{
\dontrun{

# Scenario Question: How much of symphyseal shape variance can be explained
#   by a linear regression of the cross-sectional properties of the posterior corpus?

# Posterior corpus in this example refers to the following cross-sections:

# LM1M2 - interdental junction of the left side first and second molars.
# LP3P4 - interdental junction of the left side third and fourth premolars.
# RP3P4 - interdental junction of the right side third and fourth premolars.
# RM1M2 - interdental junction of the right side first and second molars.

# Import data sample of mandibular corpus: Corpus_Land
# Import data sample of mandibular corpus data table: Biomechanics

data(Corpus_Land)
data(Biomechanics)


# Create a data analysis model

Models <- list('Regression_Model' =
                  list(
                       'Symphysis_Shape ~ LM1M2' = list(
                         'Symphysis_Shape' = Corpus_Land[241:360,,], # symphyseal landmarks
                         'LM1M2' = Biomechanics \%>\% filter(Region = "LM1M2") # LM1M2 filtered data
                         ),
                       'Symphysis_Shape ~ LP3P4' = list(
                         'Symphysis_Shape' = Corpus_Land[241:360,,],
                         'LM1M2' = Biomechanics \%>\% filter(Region = "LM1M2")
                         ),
                       'Symphysis_Shape ~ RP3P4' = list(
                         'Symphysis_Shape' = Corpus_Land[241:360,,],
                         'LM1M2' = Biomechanics \%>\% filter(Region = "LM1M2")
                         ),
                       'Symphysis_Shape ~ RM1M2' = list(
                         'Symphysis_Shape' = Corpus_Land[241:360,,],
                         'LM1M2' = Biomechanics \%>\% filter(Region = "LM1M2")
                         ),
                       )
               )


# Let's subset the filtered data tables by cross-sectional property types:

  Dimensions: Variables 1:3, 5:8, 10:15
  Orientation: Variables 9, 24
  Cortical Thickness: Variables 4, 16:18, 30:60
  Bending Resistance:Variables 19:23
  Breaking Strength: Variables 25:29


# Create the point_set argument model for subsetted variables.
# IMPORTANT: If no subsets are desired for either the response or predictor data
# all variables must be inputted in point_set subsets. For example:

point_set <- list('Regression' =
                     list(
                       'Symphysis_Shape ~ LM1M2' = list(
                         'Symphysis_Shape' = list(
                              'Subset_1' = c(1:120), # symphysis has 120 total landmarks
                              'Subset_2' = c(1:120),
                              'Subset_3' = c(1:120),
                              'Subset_4' = c(1:120),
                              'Subset_5' = c(1:120)
                              ),

                         'LM1M2' = list(
                              'Dimensions' = c(1:3, 5:8, 10:15),
                              'Orientation' = c(9, 24),
                              'Cortical Thickness' = c(4, 16:18, 30:60),
                              'Bending Resistance' = c(19:23),
                              'Breaking Strength' = c(25:29)
                              )
                         ),
                       'Symphysis_Shape ~ LP3P4' = list(
                         'Symphysis_Shape' = list(
                              'Subset_1' = c(1:120), # symphysis has 120 total landmarks
                              'Subset_2' = c(1:120),
                              'Subset_3' = c(1:120),
                              'Subset_4' = c(1:120),
                              'Subset_5' = c(1:120)
                              ),

                         'LP3P4' = list(
                              'Dimensions' = c(1:3, 5:8, 10:15),
                              'Orientation' = c(9, 24),
                              'Cortical Thickness' = c(4, 16:18, 30:60),
                              'Bending Resistance' = c(19:23),
                              'Breaking Strength' = c(25:29)
                              )
                         ),
                        'Symphysis_Shape ~ RP3P4' = list(
                         'Symphysis_Shape' = list(
                              'Subset_1' = c(1:120), # symphysis has 120 total landmarks
                              'Subset_2' = c(1:120),
                              'Subset_3' = c(1:120),
                              'Subset_4' = c(1:120),
                              'Subset_5' = c(1:120)
                              ),

                         'RP3P4' = list(
                              'Dimensions' = c(1:3, 5:8, 10:15),
                              'Orientation' = c(9, 24),
                              'Cortical Thickness' = c(4, 16:18, 30:60),
                              'Bending Resistance' = c(19:23),
                              'Breaking Strength' = c(25:29)
                              )
                         ),
                        'Symphysis_Shape ~ RM1M2' = list(
                         'Symphysis_Shape' = list(
                              'Subset_1' = c(1:120), # symphysis has 120 total landmarks
                              'Subset_2' = c(1:120),
                              'Subset_3' = c(1:120),
                              'Subset_4' = c(1:120),
                              'Subset_5' = c(1:120)
                              ),

                         'RM1M2' = list(
                              'Dimensions' = c(1:3, 5:8, 10:15),
                              'Orientation' = c(9, 24),
                              'Cortical Thickness' = c(4, 16:18, 30:60),
                              'Bending Resistance' = c(19:23),
                              'Breaking Strength' = c(25:29)
                              )
                         )
                      )
                  )

# Run the pslr.bio function to conduct parallel regressions.

output <- pslr.bio(Models, # the regression model created above
               subset = TRUE, # there are subsets in this regression test
               paired_subsets = FALSE, # Since response variable contains no unique subsets, this is not necessary
               point_set, # this is the subset model that was created above
               key = "Corpus", # Name for the output file folder.
               ) # keep everything else as default




# View the results of each regression test separately in the output object

# View the Symphysis ~ LM1M2 results

output[[1]][[1]][[1]] # structure is output, model, regression, subset
# brackets indicate first model "Regression" and the first regression test
# "Symphysis_Shape~LM1M2" and the first variable subset "Dimensions"


# view the results via the excel file listed in the folder "pslr.bio Results-Corpus":
# File name is "pslr.bio_Model_Regression.xlsx"

pslr.bio_Results = read_xlsx("pslr.bio_Model_Regression.xlsx")

view(pslr.bio_Results)
}
}
\seealso{
\code{procD.lm} \code{DT} \code{restructure_results} \code{AABA}
}
\author{
Keeling et al., 2025
}
