% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_land.R
\name{read_fcsv}
\alias{read_fcsv}
\title{Read 3D Slicer Fiducial CSV (.fcsv) Format Files}
\usage{
read_fcsv(files, dims)
}
\description{
Read 3D Slicer Fiducial CSV (.fcsv) Format Files
}
\keyword{internal}
