% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reverse.model}
\alias{reverse.model}
\title{Reverse the contents of a CSGM data model}
\usage{
reverse.model(Models)
}
\arguments{
\item{Models}{Nested list object containing regression hypotheses to test. Each list element should
contain Response and Predictor data, structured as:

\itemize{
\item 'Hypothesis Test Models' # Argument Models object name
\itemize{
\item 'Symphyseal shape ~ Posterior Corpus Properties' --- 1st hypothesis model
\itemize{
\item 'Symphyseal shape ~ LM1-M2 Corpus Properties ' --- 1st hypothesis test
\itemize{
\item Symphyseal Landmarks --- Response Data within hypothesis test
\item LM1-M2 Corpus Properties --- Predictor data within hypothesis test
\item dt_parameters --- Optional data transformation parameters
}
\item 'Symphyseal shape ~ LP3-P4 Corpus Properties ' --- 2nd hypothesis test
\itemize{
\item Symphyseal Landmarks --- Response Data within hypothesis test
\item LP3-P4 Corpus Properties --- Predictor data within hypothesis test
\item dt_parameters --- Optional data transformation parameters
}
\item 'Symphyseal shape ~ RP3-P4 Corpus Properties ' --- 3rd hypothesis test
\itemize{
\item Symphyseal Landmarks --- Response Data within hypothesis test
\item RP3-P4 Corpus Properties --- Predictor data within hypothesis test
\item dt_parameters --- Optional data transformation parameters
}
\item 'Symphyseal shape ~ RM1-M2 Corpus Properties ' --- 4th hypothesis test
\itemize{
\item Symphyseal Landmarks --- Response Data within hypothesis test
\item RM1-M2 Corpus Properties --- Predictor data within hypothesis test
\item dt_parameters --- Optional data transformation parameters
}
}
\item 'Symphyseal shape ~ Posterior Corpus Properties' --- 2nd hypothesis model
\itemize{
\item 'Symphyseal shape ~ All Corpus Properties ' --- 1st hypothesis test
\itemize{
\item Symphyseal Landmarks --- Response Data within hypothesis test
\item Posterior Corpus Properties --- Predictor data within hypothesis test
\item dt_parameters --- Optional data transformation parameters
}
}
\item ..... --- 3rd hypothesis model
}
}}
}
\value{
the same nested list object but with the predictor and response
variables reversed.
}
\description{
This function takes a nested list object following the Models argument
format within the \pkg{CSGM} package and reverses the order of the response
and predictor variables.
}
\details{
This function takes either an AABV or PSLR object that has the argument
restructure.models set to FALSE (this argument automatically performs this function).
The function is ideal when subsetted tests of variables are performed. The function
will iterate over each regression model/hypothesis and each unique regression test.
If the regression test has subsetted variables that were each separately subject
to regression tests, the function will combine each subsetted test into a single
excel sheet within an excel file for that regression model/hypothesis.
}
\examples{
\dontrun{

# Import data sample of mandibular corpus: Corpus_Land
# Import data sample of mandibular corpus data table: Biomechanics

data(Corpus_Land)
data(Biomechanics)


# Create a data analysis model

Models <- list('Regression_Model' =
                  list(
                       'Symphysis_Shape ~ LM1M2' = list(
                         'Symphysis_Shape' = Corpus_Land[241:360,,], # symphyseal landmarks
                         'LM1M2' = Biomechanics \%>\% filter(Region = "LM1M2") # LM1M2 filtered data
                         ),
                       'Symphysis_Shape ~ LP3P4' = list(
                         'Symphysis_Shape' = Corpus_Land[241:360,,],
                         'LM1M2' = Biomechanics \%>\% filter(Region = "LM1M2")
                         ),
                       'Symphysis_Shape ~ RP3P4' = list(
                         'Symphysis_Shape' = Corpus_Land[241:360,,],
                         'LM1M2' = Biomechanics \%>\% filter(Region = "LM1M2")
                         ),
                       'Symphysis_Shape ~ RM1M2' = list(
                         'Symphysis_Shape' = Corpus_Land[241:360,,],
                         'LM1M2' = Biomechanics \%>\% filter(Region = "LM1M2")
                         ),
                       )
               )

# Reverse the model

Models_revered = reverse.model(Models)
}


}
\seealso{
\code{restructure_models} \code{AABA} \code{pslr.bio}
}
\author{
Keeling et al., 2025
}
\keyword{internal}
