% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Summarize Results}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{An object resulting from either the \code{reg.bio}, \code{PSLR},
and \code{AABA} functions. Otherwise, the function uses the default \pkg{stats} package.}
}
\value{
A data table consisting of statistics for model performance.
}
\description{
A function used to extract and summarize the results from
regression based functions within the CSGM package.
}
\details{
This function extracts the regression results for the \code{GMA},
\code{PSLR} and \code{reg.bio} functions as well as conducts an internal function
\code{AABA.summary} specifically designed to extract relevant statistics for
the \code{AABA} function.
}
\seealso{
\code{\link{AABA}} \code{\link{GMA}} \code{\link{reg.bio}}
}
\author{
Keeling et al., 2025
}
