% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tune.spls}
\alias{tune.spls}
\title{tune.spls: Estimate the maximum number of variables in X to keep}
\usage{
tune.spls(X, Y, ncomp, test.keepX, ...)
}
\description{
Minimal implementation of tune.spls to satisfy dependency calls.
Returns the largest keepX to maximize information retention.
}
