% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{VIP}
\alias{VIP}
\title{Estimate Variable Importance in Projections}
\usage{
VIP(plsr.results, ncomp, Scores = FALSE, vip_threshold = "NULL")
}
\arguments{
\item{plsr.results}{a PLS object from the following PLS packages: mixOmics, pls, plsdepot}

\item{ncomp}{the number of important latent components to consider}

\item{Scores}{a logical value indicating whether or not if the VIP scores are wanted (TRUE)
or if this function should output an already trimmed predictor dataset which
surpassed the VIPs threshold}

\item{vip_threshold}{a threshold calculation method for which VIPs are found to be significant.
The general rule of thumb is for the VIPs to be greater than one which is the
a priori assumption if they were to contribute to the latent space at all. The function
calculates the vips for each meaningful component, then takes the maximum value from
all of the latent spaces. Then, vip trimming is available through vip_threshold.
Options available are the mean vip ("mean"), the median vip ("median"), and the vips > 1 ("trim").
If vip_threshold = "NULL", no variable trimming will be performed, just the scores.}
}
\description{
Use a VIP threshold to calculate variable importance in projections.
}
\details{
This function is applicable to Partial Least Squares objects from the
following packages: mixOmics, pls, plsdepot.
}
\keyword{internal}
