% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_curves.R
\name{gen_curves}
\alias{gen_curves}
\title{Generate a curve matrix for landmark data}
\usage{
gen_curves(
  ncurves,
  nfixland,
  nsemiland,
  def_points = NULL,
  def_semi = NULL,
  closed = FALSE
)
}
\arguments{
\item{ncurves}{A numeric value indicating the number of semilandmarks in your
landmark configuration.}

\item{nfixland}{A numeric value indicating the number of fixed landmarks that
should not be slid}

\item{nsemiland}{A numeric value indicating the total number of semilandmarks,
including semilandmark anchor points that form part of a curve. This value does not
include landmarks in your configuration that are unrelated to a
semilandmark curve.}

\item{def_points}{An optional vector containing numerical string representing the
landmark numbers associated with each fixed landmark point in your landmark
configuration.}

\item{def_semi}{An optional vector containing a numerical string representing the
landmark numbers associated with each semilandmark point in your landmark
configuration (not anchor points in a semilandmark configuration).
This argument is paired with def_points. Only applicable
when the landmark configurations contain fixed points independent of
the semilandmark curves.}

\item{closed}{A logical TRUE/FALSE statement indicating whether the curve
is closed in a loop.}
}
\value{
This function returns a n x 3 landmark curve matrix where n is
dependent on the number of landmarks and semilandmarks in the
landmark configuration.
}
\description{
This function generates a 'n x 3' curve matrix containing the positions of
semilandmarks relative to fixed landmarks. This curve matrix can be
subsequently used to calculate equidistant semilandmark or landmark points.
This function also satisfies the "gpagen" function's argument 'curves' from
the \emph{geomorph} package.
}
\details{
This function is primarily for landmark and semilandmark
configurations. A n x 3 landmark curve matrix is generated as a result which
can be used as an argument for functions to manage landmark spacing. This
function is only recommended for situations when the landmark configuration
contains (1) only semi-landmarks whose spacing needs further management,
(2) only landmarks that need to be equidistantly spaced, or (3) the landmark
configuration contains both semilandmarks and landmarks.
}
\examples{
\dontrun{
# You should know the exact positions of your semilandmarks in your landmark configuration before using this function

# **Simple Example 1**

# In this example, we only have 1 semilandmark curve totaling 11 semilandmarks including three fixed landmark anchor points.

Landmark_matrix <- gen_curves(ncurves = 1, # There is only 1 semilandmark curve
                              nfixland = 3, # we have 3 anchor points
                              nsemiland = 11, # we have 11 semilandmarks including the anchor points
                              def_points = NULL,
                              def_semi = NULL,
                              closed = 0 # This is an open semilandmark configuration.
                              )


 #**Simple Example 2**

# In this example, we only have 10 type 1 landmarks followed by 20 semilandmarks with two anchor points (first and last)

Landmark_matrix <- gen_curves(ncurves = 1, # There is only 1 semilandmark curve
                              nfixland = 12, # we have 2 anchor points plus 10 fixed landmarks
                              nsemiland = 20, # we have 20 semilandmarks including the anchor points
                              def_points = c(1:10, # type 1 landmarks
                                             11, 30, # Keep in mind that these are the anchor points for the open curve
                                             ),
                              def_semi = c(11:30), # You must include anchor points as well
                              closed = 0 # There are no closed semilandmarks
                              )


# **Complex Example**

# In this example, let's assume we have a landmark configuration mapping a human cranium:

\itemize{
  \item There are 32 type 1 landmarks.
  \item There are 1 *closed* semilandmark curve of 30 points mapping the right temporalis muscle attachment site.
  \item There is 1 *open* semilandmark curve of 11 points mapping the right superior temporal line.
  \item Specficially,
  \itemize{
    \item Landmarks 1-16 are type 1 landmarks.
    \item Landmarks 17 - 28 are open semilandmarks.
    \item Landmarks 29 - 39 are type 1 landmarks.
    \item Landmarks 40 - 70 are closed semilandmarks.
    \item Landmarks 71 - 75 are type 1 landmarks.
  }
}


Landmark_matrix <- gen_curves(ncurves = 2,
                              nfixland = 32,
                              nsemiland = 41,
                              def_points = c(1:16, # type 1 landmarks
                                             17, 22 ,28 # Keep in mind that these are the anchor points for the open curve
                                             29:39, # type 1 landmarks
                                             # Since the curve is close we don't include anchor points for the 2nd curve
                                             71:75 # type 1 landmarks
                                             ),
                              def_semi = c(17:28, 40:70) # You must include anchor points as well,
                              closed = 1 # There is only 1 closed semilandmark curve in this example. Default is 0.
                              )

}
}
\seealso{
\pkg{\link{geomorph}}
}
\author{
Keeling et al., 2025
}
